+++ internal/adminapi/kong.go
@@ -8,8 +8,10 @@ import (
 	"fmt"
 	"net/http"
 	"os"
+	"strings"
 
 	"github.com/kong/go-kong/kong"
+	"github.com/samber/lo"
 
 	tlsutil "github.com/kong/kubernetes-ingress-controller/v2/internal/util/tls"
 )
@@ -67,8 +69,12 @@ type HTTPClientOpts struct {
 	TLSClient TLSClientConfig
 }
 
+const (
+	headerNameAdminToken = "Kong-Admin-Token"
+)
+
 // MakeHTTPClient returns an HTTP client with the specified mTLS/headers configuration.
-func MakeHTTPClient(opts *HTTPClientOpts) (*http.Client, error) {
+func MakeHTTPClient(opts *HTTPClientOpts, kongAdminToken string) (*http.Client, error) {
 	var tlsConfig tls.Config
 
 	if opts.TLSSkipVerify {
@@ -118,8 +124,21 @@ func MakeHTTPClient(opts *HTTPClientOpts) (*http.Client, error) {
 	transport.TLSClientConfig = &tlsConfig
 	return &http.Client{
 		Transport: &HeaderRoundTripper{
-			headers: opts.Headers,
+			headers: prepareHeaders(opts.Headers, kongAdminToken),
 			rt:      transport,
 		},
 	}, nil
+}
+
+func prepareHeaders(headers []string, kongAdminToken string) []string {
+	if kongAdminToken != "" {
+		contains := lo.ContainsBy(headers, func(header string) bool {
+			return strings.HasPrefix(header, headerNameAdminToken+":")
+		})
+
+		if !contains {
+			headers = append(headers, headerNameAdminToken+":"+kongAdminToken)
+		}
+	}
+	return headers
 }